fixed up unittests
