+++ pydra/utils/__init__.py
@@ -1,1 +1,1 @@
-from .misc import user_cache_dir, add_exc_note  # noqa: F401
+from .misc import user_cache_dir, add_exc_note, exc_info_matches  # noqa: F401
