+++ pydra/utils/misc.py
@@ -1,4 +1,5 @@
 from pathlib import Path
+import re
 import platformdirs
 from pydra._version import __version__
 
@@ -31,3 +32,14 @@ def add_exc_note(e: Exception, note: str) -> Exception:
     else:
         e.args = (e.args[0] + "\n" + note,)
     return e
+
+
+def exc_info_matches(exc_info, match, regex=False):
+    if exc_info.value.__cause__ is not None:
+        msg = str(exc_info.value.__cause__)
+    else:
+        msg = str(exc_info.value)
+    if regex:
+        return re.match(".*" + match, msg)
+    else:
+        return match in msg
